% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with 15 rows and 2 columns:
\describe{
   \item{DISTANCE}{the distance the fire covered}
   \item{DAMAGE}{the amout of damage caused}
}
}
\source{
<https://people.stat.sc.edu/hitchcock/firedamageheaders.txt>
}
\usage{
fire
}
\description{
A set of data depicting fire distance and damage
}
\keyword{datasets}
