% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753JAIDENSIZEMORE.R
\docType{package}
\name{MATH4753JAIDENSIZEMORE}
\alias{MATH4753JAIDENSIZEMORE-package}
\alias{MATH4753JAIDENSIZEMORE}
\title{MATH4753JAIDENSIZEMORE Package}
\description{
This package contains functions used in MATH 4753, based on the labs from the course.
}
\author{
Jaiden Sizemore
}
