% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Make a binomial distribution plot}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{the number of iterations}

\item{n}{the size of the data set}

\item{p}{the probability of a success}
}
\value{
a barplot of the distribution
}
\description{
Make a binomial distribution plot
}
\examples{
mybin(iter = 100, n = 10, p = 0.5)

}
