% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Create a bootstrap interval for the inputted data}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{the sample}

\item{fun}{the function to be used}

\item{alpha}{the alpha value}

\item{cx}{numeric character expansion factor}

\item{...}{additional arguments for design}
}
\value{
a graph and list of useful values
}
\description{
Create a bootstrap interval for the inputted data
}
\examples{
myboot2(iter=10000, x=rchisq(20,df=3), fun="mean", alpha=0.05, xlab="mean", col="green")

}
