% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Create a 95 percent CI for mu}
\usage{
myci(x)
}
\arguments{
\item{x}{A sample}
}
\value{
A 95 percent confidence interval for the sample
}
\description{
Create a 95 percent CI for mu
}
\examples{
myci(c(1, 2, 3, 4, 5))

}
