% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Create a plot highlighting the central limit theorem's effect on Poisson distribution}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{the sample size}

\item{iter}{the number of iterations}

\item{lambda}{the lambda value of the distribution}
}
\value{
a graph of the sample distribution and a normal curve
}
\description{
Create a plot highlighting the central limit theorem's effect on Poisson distribution
}
\examples{
mycltp(5, 10000, 10)

}
