% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Create plot of max likelihood using a given function and theta}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{the fuction to be used}

\item{theta}{the theta value to be used}
}
\value{
a graph of the max likelihood and corresponding data
}
\description{
Create plot of max likelihood using a given function and theta
}
\examples{
logbin2=function(theta){log(dbinom(3,prob=theta,size=6)) + log(dbinom(5,prob=theta,size=10))}
mymaxlikg(theta = seq(0, 1, length=10000))

}
