% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Create normal distribution curve and lower tail probability}
\usage{
myncurve(x = 5, mean = 10, sd = 3)
}
\arguments{
\item{x}{the value such that P(Y < x)}

\item{mean}{the mean of the data}

\item{sd}{the standard deviation of the data}
}
\value{
a graph of the distribution and area
}
\description{
Create normal distribution curve and lower tail probability
}
\examples{
myncurve(5, 10, 3)

}
