% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Display P-value areas}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{tcalc value}

\item{xmax}{max x}

\item{n}{sample size}

\item{alpha}{alpha value}
}
\value{
Graph of pvalue range and data
}
\description{
Display P-value areas
}
\examples{
set.seed(55);x1=rnorm(30,mean=25,sd=5)
tcalc = (mean(x1)-24)/(sd(x1)/sqrt(30))
mypvalue(tcalc, n=30, alpha=0.05)

}
