% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Create discrete and continuous plots and relevant data to find the optimal tickets sold}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{the number of seats in the flight}

\item{gamma}{the probability a plane will be truly overbooked}

\item{p}{the probability of a passenger showing up}
}
\value{
Plots of discrete data, continuous data, and a list of the relevant data
}
\description{
Create discrete and continuous plots and relevant data to find the optimal tickets sold
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)

}
